+++ Assets/_Laz/Scripts/Environment/Patrol.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 906b07df815ef23429c80f46125b395b
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
