+++ Assets/_Laz/Tests/Mocks/MockWorldManager.cs
@@ -6,11 +6,20 @@ namespace Tests
     public class MockWorldManager : MonoBehaviour
     {
         [SerializeField] private LazCoordinatorBehaviour _coordinatorBehaviour = null;
-
+        [SerializeField] private PatrolBehaviour _patrolBehaviour = null;
         private void Awake()
         {
             var player = new LazPlayer();
-            _coordinatorBehaviour.Initialize(player, new ILazoWrapped[] {});
+            if (_coordinatorBehaviour != null)
+            {
+                _coordinatorBehaviour.Initialize(player, new ILazoWrapped[] { });
+            }
+
+            if (_patrolBehaviour != null)
+            {
+                _patrolBehaviour.Initialize();
+            }
+            
         }
     }
