+++ Assets/_Laz/Tests/PatrolBehaviourTests.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 957e210e28cad7f48a374f8de7ab775a
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
